refactor(api): reimplement list_functions() on the Registry
`api.py` is now the sole home for user-facing discovery functions,
including `list_functions()`.

Specifically:

- Reimplement `list_functions()` to source entries from
  `get_registry().entries` (UQTestFunInfo) instead of walking
  `UQTestFunABC` subclasses which have been removed.
- Add input_dimension, output_dimension, parameterized, and tag
  filters; add tabulate/tablefmt output modes (table, HTML, or
  plain list).
- Remove `helpers.py` entirely.
- Update the relevant tests.

Closes: #522
Ref: #544
