+++ src/uqtestfuns/global_settings.py
@@ -1,4 +1,14 @@
+from __future__ import annotations
+
 import numpy as np
 import numpy.typing as npt
 
 ARRAY_FLOAT = npt.NDArray[np.float64]
+
+SUPPORTED_TAGS = (
+    "sensitivity",
+    "optimization",
+    "metamodeling",
+    "reliability",
+    "integration",
+)
